Add attendee data to admin
