+++ wallingford_castle/admin.py
@@ -13,6 +13,39 @@ class ArcherAdmin(admin.ModelAdmin):
     autocomplete_fields = ['user', 'managing_users']
 
 
+class ArcherDataMixin(object):
+    """Utility mixin class for ModelAdmins which reference archer data."""
+
+    def archer_name(self, obj):
+        return obj.archer.name
+    archer_name.short_description = 'Name'
+
+    def archer_age(self, obj):
+        return obj.archer.get_age_display()
+    archer_age.short_description = 'Age'
+
+    def archer_agb_number(self, obj):
+        return obj.archer.agb_number
+    archer_agb_number.short_description = 'AGB number'
+
+    def archer_date_of_birth(self, obj):
+        return obj.archer.date_of_birth
+    archer_date_of_birth.short_description = 'DOB'
+
+    def archer_age_group(self, obj):
+        return obj.archer.age_group
+    archer_age_group.short_description = 'Age group'
+    archer_age_group.admin_order_field = 'archer__date_of_birth'
+
+    def archer_address(self, obj):
+        return obj.archer.address
+    archer_address.short_description = 'Address'
+
+    def archer_contact_number(self, obj):
+        return obj.archer.contact_number
+    archer_contact_number.short_description = 'Contact number'
+
+
 @admin.register(MembershipInterest)
 class MembershipInterestAdmin(DjangoObjectActions, admin.ModelAdmin):
     list_display = ['name', 'age', 'membership_type', 'status']
